lab: the store is six slices, and the two biggest panels shed their tails
lib/lab/store.ts was 800 lines with fifty-five actions in one create() call. The actions now live in lib/lab/store/<slice>.ts — project, layers, knobs, ramp, director, gallery — composed by store.ts, with the helpers more than one of them needs in store/shared.ts. Every action body is verbatim; the split was done by a script that cut the object literal at its top-level keys and computed each file's imports from what its body names, so the diff is file boundaries and import lines. LabStore, the persistence subscription and the selectors stay where they were. PixelPanel loses its two tails: the image-import modal is ImportImageModal.tsx and the undo stacks are lib/lab/pixelHistory.ts (snapshot / undo / redo / prune, the panel keeps three one-line callbacks). DirectorPanel loses its lane components (DirectorLanes.tsx) and the wire-to-pixel geometry (directorGeometry.ts, which now says BUS_WIRE_MAX instead of 1000). Verified in a browser: Director lanes render, a pen stroke on a pixel layer undoes and redoes.
