+++ web/ARCHITECTURE.md
@@ -61,7 +61,7 @@ Mirrors the firmware; the same JS runs in the browser, in the community's sandbo
 
 ## Pattern Lab (`src/app/pattern-lab/` + `src/lib/lab/`)
 
-A layered, dockable workspace: nine panels over a layer-stack project. A panel is one entry in `panels/registry.tsx` (id, title, component, where it docks by default) plus its component file — the shell derives the dockview component map, the Panels menu and the default layout from that list, and the four heavy editors load on demand. The project is one Zustand store, `src/lib/lab/store.ts`, which the render engine (`engine.ts`) reads imperatively each frame and the panels subscribe to. Everything the lab keeps in `localStorage` is listed in `src/lib/lab/persist.ts`, with the read/write helpers every persistence site uses.
+A layered, dockable workspace: nine panels over a layer-stack project. A panel is one entry in `panels/registry.tsx` (id, title, component, where it docks by default) plus its component file — the shell derives the dockview component map, the Panels menu and the default layout from that list, and the four heavy editors load on demand. The project is one Zustand store, `src/lib/lab/store.ts`, which the render engine (`engine.ts`) reads imperatively each frame and the panels subscribe to; its actions are composed from one slice per concern under `src/lib/lab/store/` (project, layers, knobs, ramp, director, gallery), with the helpers they share in `store/shared.ts`. Everything the lab keeps in `localStorage` is listed in `src/lib/lab/persist.ts`, with the read/write helpers every persistence site uses. Pixel undo stacks are `src/lib/lab/pixelHistory.ts`.
 
 | Panel | Where | What |
 | :--- | :--- | :--- |
@@ -75,7 +75,7 @@ Exports flatten the visible stack to one standalone pattern (`flatten.ts`); `hEx
 
 ## Community (`src/lib/community/` + `src/components/community/`)
 
-SQLite via Drizzle (`schema.ts`; migrations in `web/drizzle/`, applied on first open by `db.ts`), Better Auth (`auth.ts`), reads in `queries.ts`, attachments and thumbnails on disk under `communityDataDir()` (default `web/data/`, gitignored — set `COMMUNITY_DB_PATH` on a real deployment). Server-only modules (`db`, `schema`, `queries`, `auth`, `attachments`, `retention`, `builds`) sit beside `"use client"` ones (`auth-client`, `deviceHost`, `download`, `handoff`, `knobs`, `thumbs`); nothing but the file header says which is which, so check before importing one from a client component.
+SQLite via Drizzle (`server/schema.ts`; migrations in `web/drizzle/`, applied on first open by `server/db.ts`), Better Auth (`server/auth.ts`), reads in `server/queries.ts`, attachments and thumbnails on disk under `communityDataDir()` (default `web/data/`, gitignored — set `COMMUNITY_DB_PATH` on a real deployment). **`lib/community/server/` is the server side** — everything that touches SQLite, the filesystem or a session — and ESLint refuses a value import of it from `components/` or a `*Client.tsx` (type imports are fine; a value crosses through an API route). The rest of `lib/community/` is either `"use client"` (`auth-client`, `deviceHost`, `download`, `handoff`, `knobs`, `thumbs`) or pure and usable anywhere (`validate`, `license`, `deck`, `workshop`, `cors`, …).
 
 Patterns render in a sandboxed iframe (`public/pattern-sandbox.html`, driven by `components/community/SandboxPreview.tsx`; the same `pf-*` message protocol is spoken by `src/ha-card/sandbox.ts`). Pattern Lab hands off to the community through `handoff.ts` (fork vs. edit-in-place), and publishes through `components/community/PublishModal.tsx`.
